requirements.txt
Flask==2.2.5
tensorflow==2.9.1
numpy==1.23.5
pandas>=1.1.5
matplotlib>=3.3.4
scikit-learn>=0.24.2
Pillow>=8.4.0
h5py>=3.1.0
opencv-python>=4.5.5
streamlit>=1.2.0